+++ src/main/java/Funclayer/exceptions/Mapping/MapDataLayer.java
@@ -1,16 +1,17 @@
-package Funclayer.exceptions;
-
+package Funclayer.exceptions.Mapping;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
-import java.sql.SQLException;
+
+import Funclayer.exceptions.ErrorMessage;
+import Funclayer.exceptions.exceptions.DataLayerException;
 
 @Provider
 public class MapDataLayer implements ExceptionMapper<DataLayerException> {
     @Override
     public Response toResponse(DataLayerException e) {
-        ErrorMessage err = new ErrorMessage(e.getMessage(),5,"https://mama.sh/");
+        ErrorMessage err = new ErrorMessage(e.getMessage(), 5, "https://mama.sh/");
         return Response.status(Response.Status.NOT_ACCEPTABLE).entity(err).build();
     }
 }
