+++ src/main/java/Funclayer/exceptions/Mapping/MapName.java
@@ -1,14 +1,17 @@
-package Funclayer.exceptions;
+package Funclayer.exceptions.Mapping;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
+import Funclayer.exceptions.ErrorMessage;
+import Funclayer.exceptions.exceptions.NotANameException;
+
 @Provider
 public class MapName implements ExceptionMapper<NotANameException> {
     @Override
     public Response toResponse(NotANameException e) {
-        ErrorMessage err = new ErrorMessage(e.getMessage(),3,"https://mama.sh/");
+        ErrorMessage err = new ErrorMessage(e.getMessage(), 3, "https://mama.sh/");
         return Response.status(Response.Status.NOT_ACCEPTABLE).entity(err).build();
     }
 }
