+++ src/main/java/Funclayer/exceptions/exceptions/DataLayerException.java
@@ -1,4 +1,4 @@
-package Funclayer.exceptions;
+package Funclayer.exceptions.exceptions;
 
 import java.sql.SQLException;
