+++ src/main/java/Funclayer/exceptions/exceptions/UserException.java
@@ -1,4 +1,4 @@
-package Funclayer.exceptions;
+package Funclayer.exceptions.exceptions;
 
 public class UserException extends Exception {
     UserException(String message) {
