+++ src/main/java/Funclayer/exceptions/validation/CommodityValidation.java
@@ -1,0 +1,7 @@
+package Funclayer.exceptions.validation;
+
+import Funclayer.exceptions.validation.template.Validation;
+
+public class CommodityValidation extends Validation {
+
+}
