+++ src/main/java/Funclayer/exceptions/validation/PrescriptionComp.java
@@ -1,0 +1,7 @@
+package Funclayer.exceptions.validation;
+
+import Funclayer.exceptions.validation.template.Validation;
+
+public class PrescriptionComp extends Validation {
+
+}
