+++ src/main/java/Funclayer/exceptions/validation/PrescriptionValidation.java
@@ -1,0 +1,7 @@
+package Funclayer.exceptions.validation;
+
+import Funclayer.exceptions.validation.template.Validation;
+
+public class PrescriptionValidation extends Validation {
+
+}
